==> 02_scm_school/04_treated_schools_scpi_cv_result_detailed.Rmd <==
---
title: " Detailed results of cross-validated gridsearch with *scpi* package"
author: "Stefanie Meliss"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
    self_contained: true
---

This markdown summarises the results of the grid search.  

The grid search iterated through the following parameters:  

- Which measure of FSM was used? (FSM vs. FSM ever)  
- Was the outcome included as feature? (excl. DV vs. incl. DV)  
- Was any covariates adjusted for each feature? (none vs. Constant vs. Trend vs. Constant + Trend)  
- Which rolling window average width was applied to the features? (roll = 1 vs. roll = 2)  
- Was the rolling window also applied to the outcome variable? (Raw DV vs. Roll DV)
- Which years were included in the training data? (all years between 2010/11 and 2021/22)  
- Which schools were included in the donor pool? (Schools in the same region vs. schools in same and neighbouring region)  
- Was the donor pool filtered with respect to value of averaged pre-treatment outcome variable in relation to treated unit? (x SD thresholds)  
- Which weight constraints were defined during the estimation? (Simplex vs. L1-l2)  
  

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

options(scipen = 999)
# empty work space
rm(list = ls())
gc()

# load libraries
library(kableExtra)
library(dplyr)
library(data.table)
library(ggplot2)

# create function to source code
source_code <- function(root_dir_name = "code", target_repo = "helper_functions", branch = "main", file_name = "file.R") {
  
  # construct URL
  git_url <- paste0("https://raw.githubusercontent.com/stefaniemeliss/", target_repo, "/", branch, "/", file_name)
  
  # attempt to download from github
  tempp_file <- tempfile(fileext = ".R")
  
  # run processing with the parameters
  message <- tryCatch({
    curl::curl_download(git_url, tempp_file, quiet = F)
  }, error = function(e) {
    return(error = paste("Error in curl::curl_download:", e$message))
  })

  if(!grepl("Error", message)) {
    
    # if successful, source file
    source(tempp_file)
    remove(tempp_file)
    
  } else { # load local copy of file
    
    # Get the current working directory
    current_dir <- getwd()
    
    # Split the current directory into its components
    dir_components <- strsplit(current_dir, "/")[[1]]
    
    # Identify the root directory dynamically based on the provided root directory name
    root_index <- which(dir_components == root_dir_name)
    if (length(root_index) == 0) {
      stop(paste("Root directory", root_dir_name, "not found in the current path"))
    }
    root_dir <- do.call(file.path, as.list(dir_components[1:root_index]))
    
    # Identify the subdirectory one level below the root and construct its absolute path
    project_repo <- dir_components[root_index + 1]
    dir <- file.path(root_dir, project_repo)
    
    if (target_repo != project_repo) {
      dir <- gsub(project_repo, target_repo, dir) 
    }
    
    # Construct the full file path
    file_path <- file.path(dir, file_name)
    
    # Print the directory and file path for debugging
    print(paste("Directory:", dir))
    print(paste("File path:", file_path))
    
    # Source the file into the parent frame
    source(file_path, local = parent.frame())
  }
}

# source functions
source_code(target_repo = "scm_feasibility", file_name = "functions.R")

# Define the base directory
dir <- get_directory()
dir_data <- file.path(dir, "data")
dir_misc <- file.path(dir, "misc")

# get file stem name
file_stem <- get_file_stem()

# copy data #
file.copy(
  file.path(gsub("scm_feasibility", "edu_stats", dir_data), "data_swf.csv"), dir_data, overwrite = T)
file.copy(
  file.path(gsub("scm_feasibility", "edu_stats", dir_data), "data_pupils.csv"), dir_data, overwrite = T)
file.copy(
  file.path(gsub("scm_feasibility", "edu_stats", dir_data), "data_establishments_download.csv"), dir_data, overwrite = T)

# load data #

swf <- fread(file.path(dir_data, "data_swf.csv"))
pup <- fread(file.path(dir_data, "data_pupils.csv"))
est <- fread(file.path(dir_data, "data_establishments_download.csv"), na.strings = "")

# process data establishments #

# load in file with timeseries desc
summary <- read.csv(file.path(dir, "02_scm_school", "interim", "02_treated_schools_filter_donor_pool_out.csv"))

# only select schools with sufficient donor pool
summary <- subset(summary, n_pool >= 50)

# save laestab numbers
list_laestab_treated <- unique(summary$laestab)
# list_laestab_treated <- list_laestab_treated[1:5] # debug
list_laestab_treated <- list_laestab_treated[1] # debug

# create df_region as reference
df_region <- unique(summary[, c("laestab", "same", "neighbouring")])

# Create the data frame
lookup <- data.frame(time_period = c(2010:2026), 
                     time_period_str = c("2010/11", "2011/12", "2012/13", "2013/14", "2014/15", "2015/16", 
                     "2016/17", "2017/18", "2018/19", "2019/20", "2020/21", "2021/22", 
                     "2022/23", "2023/24", "2024/25", "2025/26", "2026/27"))

```



```{r plot ts, eval=FALSE, include=F}
#id_treated = list_laestab[1] # debug

# create df for plotting - treated only
# select relevant years only
tmp <- df_treat %>%
  filter(time_period %in% years) %>%
  arrange(desc(time_period)) %>%
  mutate(across(all_of(vars), ~ zoo::rollapply(.x, width = 2, FUN = function(y) mean(y, na.rm = TRUE), align = "left", partial = TRUE), .names = "{col}_roll")) %>%
  as.data.frame()

#### CREATE SPAGHETTI PLOT ####

# create data in long format
df_long <- tmp %>%
  tidyr::pivot_longer(
    cols = all_of(c(paste(vars), paste0(vars, "_roll"))),
    names_to = "variable") %>%
  mutate(category = case_match(variable, 
                               "pupil_to_qual_teacher_ratio" ~ "Outcome",
                               "pupil_to_qual_teacher_ratio_roll" ~ "Outcome",
                               "fte_avg_age_known" ~ "Teacher age",
                               "fte_avg_age_known_roll" ~ "Teacher age",
                               "pnpupfsm_e" ~ "% pupils FSM",
                               "pnpupfsm_e_roll" ~ "% pupils FSM",
                               "pnpupfsm_ever" ~ "% pupils FSM ever",
                               "pnpupfsm_ever_roll" ~ "% pupils FSM ever"
  ),
  computation = ifelse(grepl("roll", variable), "Rolling average", "Raw")) %>%
  as.data.frame()

df_long$category <- factor(df_long$category, levels = c("Outcome", "% pupils FSM", "% pupils FSM ever", "Teacher age"))


# determine time period
period.avail.plt <- sort(unique(df_long$time_period))
period.pre.plt <- setdiff(period.avail.plt, c(2022:2023)) # Pre-treatment period

# spaghetti plot
plt <- ggplot(data = df_long, aes(x = time_period, y = value, col = computation, group = interaction(laestab, computation))) +
  geom_vline(xintercept = period.pre.plt[length(period.pre.plt)]+0.5, linetype = "dotted") +
  geom_line() + 
  geom_point() +
  facet_wrap(~ category, ncol = 1, strip.position = "top", scales = "free_y") +
  ambition_theme +
  ylab("Reported value") + xlab("Academic year") +
  scale_x_continuous(breaks = df_long$time_period, labels = df_long$time_period_str) +
  scale_color_manual(values = c("Raw" = blue, "Rolling average" = red)) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        legend.title = element_blank())

suppressWarnings(print(plt))
cat("\n\n")  # Add some space between plots

```

```{r desc ts, eval=FALSE, include=F}

# -- get timeseries descriptives -- #

# compute timeseries descriptives
ts_desc <- apply(tmp[, grepl("pup|fte", names(tmp))], MARGIN = 2, FUN = function(x){psych::describe(x, IQR = T, trim = 1/nrow(tmp))})

# combine to df
ts_desc <- do.call("rbind",ts_desc)

# add period
ts_desc$period <- period

# select columns
ts_desc <- ts_desc[, c("period", "n", "mean", "sd", "min", "max")]

# compute timeseries auto correlation
max_lag = ceiling(.25 * nrow(tmp)) # use lags up to about one-quarter of the total number of observations.
ts_ac <- apply(tmp[, grepl("pup|fte", names(tmp))], MARGIN = 2, FUN = function(x){
  x <- ifelse(is.na(x), 0, x)
  tmp <- acf(ts(na.omit(x)), lag.max = max_lag, plot = F)
  return(tmp$acf)
})

# transpose output
ts_ac <- as.data.frame(t(ts_ac))
names(ts_ac) <- paste0("ac_l", 0:(ncol(ts_ac)-1))
ts_ac$ac_l0 <- NULL

# combine all measures
out <- merge(ts_desc, ts_ac, by = 0)
names(out)[1] <- "Variable"

# assign new variable names
out$Variable <- factor(out$Variable, 
                       levels = c("pupil_to_qual_teacher_ratio", "pupil_to_qual_teacher_ratio_roll", 
                                  "pnpupfsm_e", "pnpupfsm_e_roll",
                                  "pnpupfsm_ever", "pnpupfsm_ever_roll",
                                  "fte_avg_age_known", "fte_avg_age_known_roll"
                       ),
                       labels = c("Outcome (raw)", "Outcome (roll)",
                                  "FSM (raw)", "FSM (roll)",
                                  "FSM ever (raw)", "FSM ever (roll)",
                                  "Age (raw)", "Age (roll)")
                       
)
# sort 
out <- out[order(out$Variable), ]

# print to markdown
kbl(out[, -2], caption = paste0("Descriptives of timeseries data (", period, ")"), digits = 2, row.names = F) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed"), fixed_thead = T) %>% 
  print()
cat("\n")


```



```{r facets, eval=FALSE, include=F}

# --- create new vars for facets --- #

# was rolling average computed
results$rolling.window <- ifelse(results$rolling.window == 1, "Raw IV", "Roll IV")
results$rolling.window <- factor(results$rolling.window)

# was rolling average computed for outcome
results$roll.outcome <- ifelse(results$roll.outcome, "Roll DV", "Raw DV")
results$roll.outcome <- factor(results$roll.outcome)

# donor pool filtering options # 

# which region was used for filtering
results$region.filter <- ifelse(grepl("neighbouring", results$region.filter), "Same + Neighb.", "Same")
results$region.filter <- factor(results$region.filter)

# which SD filter was applied?
if ("50" %in% unique(results$sd.range)) {
  # Primary
  results$sd.range <- ifelse(grepl("50", results$sd.range), "0.5 SD", "1 SD")
  results$sd.range <- factor(results$sd.range)
} else {
  # Secondary
  results$sd.range <- ifelse(grepl("100", results$sd.range), "1 SD", "No SD")
  results$sd.range <- factor(results$sd.range, levels = c("No SD", "1 SD"))
}

# which years were used in training
results$year.min <- sapply(results$period.pre, function(x) extract_min_max_years(x)[1])
results$year.max <- sapply(results$period.pre, function(x) extract_min_max_years(x)[2])
results$period.pre <- paste0(results$year.min, ":", results$year.max)
results$period.pre <- factor(results$period.pre)

if (scm_setup) {
  # SCM setup options in scpi #
  
  # which FSM measure?
  results$fsm <- ifelse(grepl("pnpupfsm_ever", results$features), "FSM ever", "FSM")
  results$fsm <- factor(results$fsm)
  
  # is DV in features?
  results$feat_dv <- ifelse(grepl("ratio", results$features), "incl. DV", "excl. DV")
  results$feat_dv <- factor(results$feat_dv)
  
  # covariate adjustment
  results$adj <- ifelse(is.na(results$cov.adj), "none", results$cov.adj)
  results$adj <- factor(results$adj, levels = c("none", "constant", "trend", "constant, trend"),
                            labels = c("none", "Constant", "Trend", "C + T"))
  
  # which constraints were applied to the weights
  results$method <- tstrsplit(results$w.constr, "; ")[[1]]
  results$method <- gsub("name = ", "", results$method)
  results$method <- factor(results$method, levels = c("simplex", "L1-L2", "lasso", "user provided"))
  
  # create a bool capturing if there was an error
  results$error <- grepl("Error", results$status_it)
  
}

```



```{r proc grid donor, eval=FALSE, echo=F}
# read in results
file_name <- file.path(dir, "02_scm_school", "interim", paste0("03_treated_schools_scpi_cv_", gsub(" ", "_", id_name), "_ts_donor.csv"))
results <- read.csv(file = file_name)

# PROCESS RESULTS #
scm_setup <- FALSE
<<facets>>

# focus on RAW data
results_donor <- results[results$rolling.window == "Raw IV" & results$roll.outcome == "Raw DV", ]
rm(results)

# overwrite labels of some facets for clarity
results_donor$sd.range <- paste0(results_donor$sd.range, " ( = ", results_donor$n_pool, " donors)")
results_donor$region.filter <- paste0("Region: ", results_donor$region.filter)

# create data in long format
df_long <- results_donor %>%
  filter(time_period <= 2023) %>%
  tidyr::pivot_longer(
    cols = all_of(vars),
    names_to = "variable") %>%
  mutate(category = case_match(variable, 
                               "pupil_to_qual_teacher_ratio" ~ "Outcome",
                               "fte_avg_age_known" ~ "Teacher age",
                               "pnpupfsm_e" ~ "% pupils FSM",
                               "pnpupfsm_ever" ~ "% pupils FSM ever")) %>%
  as.data.frame()

df_long$category <- factor(df_long$category, levels = c("Outcome", "% pupils FSM", "% pupils FSM ever", "Teacher age"))

```

```{r plot grid donor, eval=FALSE, echo=F}
cat("### Effects on the donor pool\n\n")

tmpp <- df_long %>% 
  filter(period.pre == period)

# plot timeseries average for each school
cols <- c(navy40, coral)
names(cols) <- c("Donor school", id_name)

for(region.filter in sort(unique(tmpp$region.filter))){

    tmp <- tmpp[tmpp$region.filter == region.filter, ]
    
    plt <- ggplot(data = tmp, aes(x = time_period, y = value, col = status, group = laestab)) +
      geom_line(data = tmp[tmp$status == "Donor school", ], aes(col = paste("Donor school"))) + 
      geom_line(data = tmp[tmp$status == paste(id_name), ], aes(col = paste(id_name)), linewidth =.8) +
      geom_point(data = tmp[tmp$status == paste(id_name), ], aes(col = paste(id_name)), size = 2) +
      ggh4x::facet_grid2(cols = vars(sd.range), rows = vars(region.filter, category), strip = ggh4x::strip_nested(), scales = "free_y") +
      ambition_theme +
      theme(strip.text = element_text(size = 10),
            legend.title = element_blank(),
            axis.title = element_text(size = 10),
            axis.text = element_text(size = 10),
            axis.text.x = element_text(angle = 45, hjust = 1)
      ) +
      scale_x_continuous(breaks = tmp$time_period, labels = tmp$time_period_str) +
      scale_color_manual(
        breaks=c(id_name, "Donor school"),
        values=cols) +
      ylab("Reported value") + xlab("Academic year")
    
    print(plt)
    cat("\n\n")
    
    rm(tmp)
    
}


```



```{r proc grid fit, eval=FALSE, echo=F}

#### process data grid search ####

# declare file name
file_name <- file.path(dir, "02_scm_school", "interim", paste0("03_treated_schools_scpi_cv_", gsub(" ", "_", id_name), "_results.csv"))

# read in results
results <- read.csv(file_name)

# PROCESS RESULTS #
scm_setup <- TRUE
<<facets>>


# --- process data for plotting --- #

# identify promising combinations
results_fit <- results %>%
  # remove all error runs
  filter(!error) %>%
  # create one col per rmspe
  mutate(
    sd_all = sd_treated[cross.val == F],
    rmspe_all = rmspe_pre[cross.val == F], 
    rmspe_sim = rmspe_post[cross.val == F], 
    rmspe_T = rmspe_pre[cross.val == T], 
    rmspe_V = rmspe_post[cross.val == T],
    .by = it
  ) %>%
  # compute ratios #
  # if ratio is below 1: worse fit in training than in validation data
  # if ratio is above 1: worse fit in training than in validation data
  # if ratio is equal to 1: equal fit in training and validation data
  mutate(ratio = rmspe_V / rmspe_T) %>%
  # run checks
  mutate(
    # compare all rmspe against pretreatment SD
    rmspe_below_sd = rmspe_all < sd_all & rmspe_T < sd_all & rmspe_V < sd_all,
    # compare ratio against limits
    ratio_within_limits = between(ratio, .8, 1.2),
    # bookmark good parameter combinations
    promising = as.logical(rmspe_below_sd * ratio_within_limits))


# convert to long format
tmp_long <- results_fit %>%
  # from wide to long
  tidyr::pivot_longer(
    cols = c(rmspe_pre, rmspe_post),
    names_to = "data",
    values_to = "rmspe") %>%
  # remove all rows for rmspe in CV runs
  filter(!(cross.val == F & data == "rmspe_post")) %>%
  # rename
  mutate(data = case_match(data, 
                           "rmspe_pre" ~ "T",
                           "rmspe_post" ~ "V"  )) %>%
  mutate(data = ifelse(cross.val == F, "all", data)) %>%
  mutate(data = factor(data, levels = c("T", "V", "all"))) %>%
  # Explicitly arrange dataset by all relevant variables
  arrange(it, feat_dv, adj, fsm, region.filter, sd.range, rolling.window, roll.outcome, period.pre, method, data)


# Define explicit numeric dodge positions
method_levels <- unique(results_fit$method)
dodge_width <- 0.6
n_methods <- length(method_levels)
# Create evenly spaced numeric offsets around 0
method_positions <- seq(-dodge_width/2, dodge_width/2, length.out = n_methods)

# Create a lookup table
position_lookup <- tibble(
  method = method_levels,
  method_position = method_positions
) %>%
  mutate(method = factor(method, levels = c("simplex", "L1-L2", "lasso", "user provided"))
  )

# Merge these positions into your data
tmp_long <- tmp_long %>%
  left_join(position_lookup, by = "method") %>%
  mutate(
    data_numeric = as.numeric(factor(data)), # numeric x-axis position per 'data'
    x_position = data_numeric + method_position # final numeric position for plotting
  )

```

```{r plot grid fit, eval=FALSE, echo=F}
pos_y = max(tmp_long$rmspe)+1

#### plot data grid search ####

cat("### Effects on the model fit\n\n")

tmpp <- tmp_long %>% # derivative of results_fit
      filter(year.min == min(years) & !error)


for(region.filter in sort(unique(tmpp$region.filter))){
  
  for(sd.range in sort(unique(tmpp$sd.range))){

    # subset data for plotting
    tmp <- tmpp[tmpp$region.filter == region.filter & tmpp$sd.range == sd.range,]
    
    plt <- ggplot(data = tmp, 
                  aes(x = x_position, y = rmspe, group = it)) +
      geom_hline(data = function(y) y %>% filter(data == "all") %>% group_by(feat_dv, adj, region.filter, fsm, sd.range, rolling.window, roll.outcome) %>% summarise(sd_treated = unique(sd_treated), .groups = 'drop'),
                 aes(yintercept = sd_treated), color = black40, linetype = "dashed") +
      geom_point(aes(colour = method), size = 1) +
      geom_line(aes(colour = method)) +
      ggh4x::facet_grid2(cols = vars(region.filter, sd.range, fsm, rolling.window, roll.outcome), rows = vars(feat_dv, adj), strip = ggh4x::strip_nested()) +
      geom_label(data = tmp[tmp$data == "all", ], aes(x = 2, y = pos_y, label = paste0("it ", it,": ", n_pool, "|", n_active), fill =  rmspe_below_sd), 
                 hjust = 0.5, vjust = 0.5, show.legend = F) +
      scale_x_continuous(
        breaks = unique(tmp$data_numeric),
        labels = unique(tmp$data)
      ) +
      ambition_theme +
      theme(strip.text = element_text(size = 10),
            legend.title = element_text(size = 10),
            legend.text = element_text(size = 10),
            axis.title = element_text(size = 10),
            axis.text = element_text(size = 10),
            #axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
            plot.caption = element_text(size = 10),
            legend.margin=margin(t = -8)) +
      ylab("Root Mean Squared Prediction Error (RMSPE)") +
      xlab("Cross-validation fold") +
      guides(colour = guide_legend(title = "Weight constraint")) +
      scale_colour_manual(values = ambition_palette_bright) +
      scale_fill_manual(values= c("TRUE" = navy40, "FALSE" = white))

    print(plt)
    cat("\n\n")
    
  }
}


```



```{r proc grid gap, eval=FALSE, echo=F}
#### process data grid search ####

# declare file name
file_name <- file.path(dir, "02_scm_school", "interim", paste0("03_treated_schools_scpi_cv_", gsub(" ", "_", id_name), "_ts_synth.csv"))

# read in results
results <- read.csv(file_name)

# PROCESS RESULTS #
scm_setup <- TRUE
<<facets>>


# identify promising combinations
results_gap <- results %>%
    # remove all error runs
  filter(!error) %>%
  # create one col per rmspe
  mutate(
    sd_all = sd_treated[cross.val == F & time_period == 2021],
    rmspe_all = rmspe_pre[cross.val == F & time_period == 2021], 
    rmspe_sim = rmspe_post[cross.val == F & time_period == 2021], 
    rmspe_T = rmspe_pre[cross.val == T & time_period == 2021], 
    rmspe_V = rmspe_post[cross.val == T & time_period == 2021],
    .by = it
  ) %>%
  # compute ratios #
  # if ratio is below 1: worse fit in training than in validation data
  # if ratio is above 1: worse fit in training than in validation data
  # if ratio is equal to 1: equal fit in training and validation data
  mutate(ratio = rmspe_V / rmspe_T) %>%
  # run checks
  mutate(
    # compare all rmspe against pretreatment SD
    rmspe_below_sd = rmspe_all < sd_all & rmspe_T < sd_all & rmspe_V < sd_all,
    # compare ratio against limits
    ratio_within_limits = between(ratio, .8, 1.2),
    # bookmark good parameter combinations
    promising = as.logical(rmspe_below_sd * ratio_within_limits))


# focus on SCM estimations
results_gap <- results_gap[results_gap$method == "simplex", ]

```

```{r plot grid gap, eval=FALSE, echo=F}
tmpp <- results_gap %>% 
  filter(year.min == min(years) & !error)

#### plot data grid search ####

cat("### Effects on the synthetic timeseries \n\n")

cols <- c(red, blue, black)
names(cols) <- c("Synthetic school CV", "Synthetic school", id_name)


xintercept1 = eval(parse(text = period))[length(eval(parse(text = period)))-2]+0.5
xintercept2 = eval(parse(text = period))[length(eval(parse(text = period)))]+0.5
x_mid = eval(parse(text = period))[ceiling((length(eval(parse(text = period)))+3)/2)]

for(region.filter in sort(unique(tmpp$region.filter))){
  
  for(sd.range in sort(unique(tmpp$sd.range))){
    
    # subset data for plotting
    tmp <- tmpp[tmpp$region.filter == region.filter & tmpp$sd.range == sd.range,]
    
    # plot treated against synthetic unit
    plt <- ggplot() +
      geom_vline(xintercept = xintercept1, linetype = "dotted") +
      geom_vline(xintercept = xintercept2, linetype = "dotted") +
      geom_label(data = tmp[tmp$cross.val == F, ], aes(x = x_mid, y = 25.5, label = paste0("it ", it,": ", n_pool, "|", n_active), fill = rmspe_below_sd),
                 hjust = 0.5, vjust = 0.5, show.legend = F) +
      geom_line(data = tmp, aes(x = time_period, y = actual, col = paste(id_name))) +
      geom_line(data = tmp[tmp$cross.val == T, ], aes(x = time_period, y = synthetic, col = "Synthetic school CV")) +
      geom_line(data = tmp[tmp$cross.val == F, ], aes(x = time_period, y = synthetic, col = "Synthetic school")) +
      ggh4x::facet_grid2(cols = vars(region.filter, sd.range, fsm, rolling.window, roll.outcome), rows = vars(feat_dv, adj), strip = ggh4x::strip_nested()) +
      ambition_theme +
      theme(axis.text.x = element_text(angle = 45, hjust = 1),
            legend.title = element_blank(),
            strip.text = element_text(size = 10),
            legend.text = element_text(size = 10),
            axis.title = element_text(size = 10),
            axis.text = element_text(size = 10),
            plot.caption = element_text(size = 10),
            legend.margin=margin(t = -8)) +
      coord_cartesian(ylim = c(12, 26)) +
      ylab("Ratio of pupils to qualified teachers") + xlab("Academic year") +
      scale_x_continuous(breaks = lookup$time_period[seq(1, length(lookup$time_period), 2)], labels = lookup$time_period_str[seq(1, length(lookup$time_period_str), 2)]) +
      scale_color_manual(breaks = c(id_name, "Synthetic school CV", "Synthetic school"), values = cols) +
      scale_fill_manual(values= c("TRUE" = navy40, "FALSE" = white))

    print(plt)
    cat("\n\n")

    
    # plot gap
    plt <- ggplot() +
      geom_vline(xintercept = xintercept1, linetype = "dotted") +
      geom_vline(xintercept = xintercept2, linetype = "dotted") +
      geom_hline(yintercept = 0, linetype = "dashed") +
      geom_label(data = tmp[tmp$cross.val == F, ], aes(x = x_mid, y = 4.5, label = paste0("it ", it,": ", n_pool, "|", n_active), fill = rmspe_below_sd),
                 hjust = 0.5, vjust = 0.5, show.legend = F) +
      geom_line(data = tmp[tmp$cross.val == T, ], aes(x = time_period, y = gap, col = "Synthetic school CV")) +
      geom_line(data = tmp[tmp$cross.val == F, ], aes(x = time_period, y = gap, col = "Synthetic school")) +
      ggh4x::facet_grid2(cols = vars(region.filter, sd.range, fsm, rolling.window, roll.outcome), rows = vars(feat_dv, adj), strip = ggh4x::strip_nested()) +
      ambition_theme +
      theme(axis.text.x = element_text(angle = 45, hjust = 1),
            legend.title = element_blank(),
            strip.text = element_text(size = 10),
            legend.text = element_text(size = 10),
            axis.title = element_text(size = 10),
            axis.text = element_text(size = 10),
            plot.caption = element_text(size = 10),
            legend.margin=margin(t = -8)) +
      coord_cartesian(ylim = c(-5, 5)) +
      ylab("Gap in ratio of pupils to qualified teachers") + xlab("Academic year") +
      scale_x_continuous(breaks = lookup$time_period[seq(1, length(lookup$time_period), 2)], labels = lookup$time_period_str[seq(1, length(lookup$time_period_str), 2)]) +
      scale_color_manual(breaks = c(id_name, "Synthetic school CV", "Synthetic school"), values = cols) +
      scale_fill_manual(values= c("TRUE" = navy40, "FALSE" = white))

    print(plt)
    cat("\n\n")
    
    
    
    # show promising combinations
    tmp <- tmp %>%
      # prepare for printing
      filter(time_period == 2021 &
               method == "simplex" &
               !error &
               !cross.val) %>%
      select(c(rmspe_below_sd, it, feat_dv, adj, region.filter, sd.range, fsm, rolling.window, roll.outcome,
               n_pool, n_active, sd_all, rmspe_all, rmspe_T, rmspe_V, ratio, m_gap, sd_gap, min_gap, max_gap)) %>%
      arrange(feat_dv, adj, fsm, rolling.window, roll.outcome) 

    row_idx <- which(tmp$rmspe_below_sd == T)
    # print #
    kbl(tmp, row.names = F, digits = 3) %>%
      kable_styling(bootstrap_options = c("striped", "hover", "condensed"), fixed_thead = T) %>%
      column_spec(c(1,2), bold = T) %>%
      column_spec(c(2), border_right =  T) %>%
      row_spec(row_idx, bold = T) %>%
      print()
    cat("\n\n")
  }
}

```



```{r, echo=F, message=FALSE, results='asis', fig.align='center', fig.height=10, fig.width=13, out.width='100%', out.height='100%'}
i = 1 # debug
for (i in 1:length(list_laestab_treated)) {
  
  id_treated <- list_laestab_treated[i]
  
  # define options for filtering for each school
  
  if (id_treated == 3804004) { # Dixons Kings Academy
    swf_filter = "! time_period %in% c(201112, 201213, 201314)" # gap in data for 2013/14
  } else {
    swf_filter = NULL
  }
  
  # process data
  data <- process_data_scm(id_treated = id_treated,
                           var_teach = c("fte_avg_age_known"),
                           var_pup = c("pnpupfsm_e", "pnpupfsm_ever"),
                           swf_filter = swf_filter)

  cat("# ", id_name, "\n\n")
  
  # print some information about the school
  cat("Type of establishment:", est_treated$typeofestablishment_name, "\n\n")
  cat("Phase:", est_treated$phaseofeducation_name, "\n\n")
  cat("Gender of pupils:", est_treated$gender_name, "\n\n")
  cat("Religious character:", est_treated$religiouscharacter_name, "\n\n")
  cat("Trust flag:", est_treated$trustschoolflag_name, "\n\n")
  cat("Local authority:", est_treated$la_name, "\n\n")
  cat("Region:", est_treated$gor_name, "\n\n")
  
  
# process donor grid data
<<proc grid donor>>

# process fit grid data
<<proc grid fit>>
    
# process gap grid data
<<proc grid gap>>

  # load in grid #
  # declare file name
  file_name <- file.path(dir, "02_scm_school", "interim", paste0("03_treated_schools_scpi_cv_", gsub(" ", "_", id_name), "_grid.csv"))
  
  # read in results
  param_grid <- read.csv(file_name)
  
  # get different period.pre information
  periods <- unique(param_grid$period.pre[param_grid$cross.val == F])
  
  
  for (period in periods) {
    
    cat("## Pre-treatment timeseries", period, "\n\n")
    
    # determine years based on character string input
    years <- eval(parse(text = period))
    
# Plot timeseries treated unit
<<plot ts>>
# Compute descriptives of timeseries for treated unit
<<desc ts>>
# Show donor pool
<<plot grid donor>>

# Show fit
<<plot grid fit>>

# Show synthetic timeseries
<<plot grid gap>>
   
    
  }
  
}

```
